fix2

+++ hellobindings/src/terrain_3d/geoclipmap.rs
@@ -314,23 +314,23 @@ impl GeoClipMap {
             indices.resize(tile_resolution * 24 + 6);
             n = 0;
             let vertices_mut = vertices.as_mut_slice();
-            for i in 0..(patch_vert_resolution * 2) {
-                vertices_mut[n] = Vector3::new((i - tile_resolution) as f32, 0.0, 0.0);
+            for i in 0..(patch_vert_resolution * 2) as i32 {
+                vertices_mut[n] = Vector3::new((i - tile_resolution as i32) as f32, 0.0, 0.0);
                 aabb.expand(vertices_mut[n]);
                 n += 1;
 
-                vertices_mut[n] = Vector3::new((i - tile_resolution) as f32, 0.0, 1.0);
+                vertices_mut[n] = Vector3::new((i - tile_resolution as i32) as f32, 0.0, 1.0);
                 aabb.expand(vertices_mut[n]);
                 n += 1;
             }
 
             let start_of_vertical = n as i32;
-            for i in 0..(patch_vert_resolution * 2) {
-                vertices_mut[n] = Vector3::new(0.0, 0.0, (i - tile_resolution) as f32);
+            for i in 0..(patch_vert_resolution * 2)  as i32 {
+                vertices_mut[n] = Vector3::new(0.0, 0.0, (i - tile_resolution as i32) as f32);
                 aabb.expand(vertices_mut[n]);
                 n += 1;
 
-                vertices_mut[n] = Vector3::new(1.0, 0.0, (i - tile_resolution) as f32);
+                vertices_mut[n] = Vector3::new(1.0, 0.0, (i - tile_resolution as i32) as f32);
                 aabb.expand(vertices_mut[n]);
                 n += 1;
             }
@@ -394,25 +394,25 @@ impl GeoClipMap {
             indices.resize(clipmap_vert_resolution * 6);
             n = 0;
             let vertices_mut = vertices.as_mut_slice();
-            for i in 0..clipmap_vert_resolution {
-                n = clipmap_resolution * 0 + i;
+            for i in 0..clipmap_vert_resolution as i32 {
+                n = clipmap_resolution * 0 + i as usize;
                 vertices_mut[n] = Vector3::new(i as f32, 0.0, 0.0);
                 aabb.expand(vertices_mut[n]);
 
-                n = clipmap_resolution * 1 + i;
+                n = clipmap_resolution * 1 + i as usize;
                 vertices_mut[n] = Vector3::new(clipmap_vert_resolution as f32, 0.0, i as f32);
                 aabb.expand(vertices_mut[n]);
 
-                n = clipmap_resolution * 2 + i;
+                n = clipmap_resolution * 2 + i as usize;
                 vertices_mut[n] = Vector3::new(
-                    (clipmap_vert_resolution - i) as f32,
+                    (clipmap_vert_resolution as i32 - i) as f32,
                     0.0,
                     clipmap_vert_resolution as f32,
                 );
                 aabb.expand(vertices_mut[n]);
 
-                n = clipmap_resolution * 3 + i;
-                vertices_mut[n] = Vector3::new(0.0, 0.0, (clipmap_vert_resolution - i) as f32);
+                n = clipmap_resolution * 3 + i as usize;
+                vertices_mut[n] = Vector3::new(0.0, 0.0, (clipmap_vert_resolution as i32 - i) as f32);
                 aabb.expand(vertices_mut[n]);
             }
